Add files via upload
This Responsive Parallax Flipping Cards are a common design concept that can be accomplished with HTML and CSS. The main concept is to construct a card-like structure that, when clicked or hovered over, flips or rotates, revealing more material on the reverse of the card.

This effect is created by combining CSS 3D transformations and transitions. The HTML structure is commonly made up of a container element and two child elements, one for the front and one for the back of the card.

The parallax and flipping effects are created by applying CSS styles to these components. here we ensure that the cards are responsive and adjust to different screen sizes, for that CSS media queries can be used to adjust the styles for different viewport widths.

HTML Structure

Let's start with making its HTML structure, Firstly it creates a container with four columns, each containing a card with a front and back. The card's front contains paragraph and some text, while the back has a longer.

Next the div tag with the class wrapper is the outermost container that wraps around all the columns. Inside this div tag is another div tag with the class cols, which contains the four columns. Each and every column is created using a div tag with the class col. The on touch start attribute specifies a JavaScript function that will be called when the user touches the card, toggling the hover class on and off.
The front of the card is created using a div tag with the class front, which has a background-image property that is currently set to an empty URL. You can put any image in it, it will not make any difference to the shape of the card.

The back of the card is created using a div tag with the class back, Inside the div tag with the class back is another div tag with the class inner. The inner div contains a single paragraph element (p) with Lorem Ipsum text. In this you can describe something or you can also write about someone.

When the user hovers over the front of the card, the hover class is toggled on the parent col div using the ontouchstart JavaScript function specified on each col element. This will cause the card to flip over, revealing the back of the card with the longer description.


Styling With CSS

Till now we have made only the structure of these cards, now comes the most important work. it's time to design these cards with CSS.First of all we will start with designing the wrapper class. the .wrapper sets the width and max-width of the component and centers it horizontally using the margin property. the width of the container to 90% of its parent element's width and it also sets the top and bottom margins to 0.

next comes the .cols it is a class selector that applies to a group of elements that will be displayed in a flexbox layout. in this the display mode of the element to a flexible container, so that its child elements can be arranged using flexbox properties.
lastly .col is a class selector that applies to each of the child elements within the .cols container. the width Sets the width of each column to 25% of the width of the .cols container, minus 2rem. The calc() function is used to subtract 2rem from the width calculation.

And .front and .back style the front and back face of the card, respectively. They set the background, border-radius, height, and font-size of the card. They also use transitions to animate the transform property when the card is hovered.

.front:after creates a semi-transparent overlay over the front face of the card. and .inner styles the content inside the card, setting its position and padding.

.container .back and .container .front use transforms to rotate the back and front faces 180 degrees around the Y-axis, making them invisible when they're not supposed to be shown.

.front .inner p styles the text content of the card by setting the font-size, margin-bottom, and font-family of the paragraph element.